+++ src/main/java/kr/hhplus/be/server/common/event/Event.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.common.event;
+
+public interface Event {
+}
