+++ src/main/java/kr/hhplus/be/server/common/event/EventPublisher.java
@@ -1,5 +1,5 @@
 package kr.hhplus.be.server.common.event;
 
-public interface EventPublisher {
-    void publish(Object event);
+public interface EventPublisher<E extends Event> {
+    void publish(E event);
 }
