--- src/main/java/kr/hhplus/be/server/common/event/infra/SpringEventPublisher.java
@@ -1,18 +1,0 @@
-package kr.hhplus.be.server.common.event.infra;
-
-import kr.hhplus.be.server.common.event.EventPublisher;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.stereotype.Component;
-
-@RequiredArgsConstructor
-@Component
-public class SpringEventPublisher implements EventPublisher {
-
-    private final ApplicationEventPublisher applicationEventPublisher;
-
-    @Override
-    public void publish(Object event) {
-        applicationEventPublisher.publishEvent(event);
-    }
-}
