+++ src/main/java/kr/hhplus/be/server/order/application/event/DataPlatformEventHandler.java
@@ -1,5 +1,6 @@
 package kr.hhplus.be.server.order.application.event;
 
+import kr.hhplus.be.server.order.domain.event.OrderCreatedEvent;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
