--- src/main/java/kr/hhplus/be/server/order/application/event/OrderCreatedProduct.java
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.order.application.event;
-
-public record OrderCreatedProduct(
-        long productId,
-        int quantity
-) {
-}
