+++ src/main/java/kr/hhplus/be/server/order/application/facade/OrderFacade.java
@@ -1,13 +1,12 @@
 package kr.hhplus.be.server.order.application.facade;
 
-import kr.hhplus.be.server.common.event.EventPublisher;
 import kr.hhplus.be.server.common.exception.BusinessException;
 import kr.hhplus.be.server.common.exception.ErrorCode;
 import kr.hhplus.be.server.common.lock.DistributedLock;
 import kr.hhplus.be.server.coupon.application.service.CouponService;
 import kr.hhplus.be.server.order.application.command.OrderCreateCommand;
-import kr.hhplus.be.server.order.application.event.OrderCreatedEvent;
-import kr.hhplus.be.server.order.application.event.OrderCreatedProduct;
+import kr.hhplus.be.server.order.domain.event.OrderCreatedEvent;
+import kr.hhplus.be.server.order.application.event.OrderEventPublisher;
 import kr.hhplus.be.server.order.application.result.OrderAggregate;
 import kr.hhplus.be.server.order.application.result.OrderResult;
 import kr.hhplus.be.server.order.application.service.OrderService;
@@ -29,7 +28,7 @@ public class OrderFacade {
     private final ProductService productService;
     private final CouponService couponService;
     private final PointService pointService;
-    private final EventPublisher eventPublisher;
+    private final OrderEventPublisher orderEventPublisher;
 
     @DistributedLock(
             keys = {
@@ -56,10 +55,10 @@ public class OrderFacade {
             throw new BusinessException(ErrorCode.CONFLICT_USE);
         }
 
-        eventPublisher.publish(new OrderCreatedEvent(
+        orderEventPublisher.publish(new OrderCreatedEvent(
                 orderedAt,
                 command.products().stream()
-                        .map(product -> new OrderCreatedProduct(product.productId(), product.quantity()))
+                        .map(product -> new OrderCreatedEvent.OrderProduct(product.productId(), product.quantity()))
                         .toList()
         ));
