+++ src/main/java/kr/hhplus/be/server/order/domain/event/OrderEvent.java
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.order.domain.event;
+
+import kr.hhplus.be.server.common.event.Event;
+
+public interface OrderEvent extends Event {
+}
